separate all components and images by domain
