+++ src/hooks/main/useCategories.ts
@@ -1,0 +1,5 @@
+import { useAPIQuery } from "../common/useAPIQuery";
+
+export function useCategories() {
+  return useAPIQuery("getAllCategories", { count: 12 }, { suspense: true }).data.data;
+}
