+++ src/routes/RouteSwitch.tsx
@@ -1,7 +1,7 @@
 import { Suspense } from "react";
 import { Switch, Route } from "react-router-dom";
 
-import { useScrollToTop } from "../hooks/useScrollToTop";
+import { useScrollToTop } from "../hooks/common/useScrollToTop";
 
 import { routes, RouteSpec } from "./routeSpec";
